a more convinient way to specify uploader
The `micronuclus` bootloaer/uploader are famous for several ATtiny
MCUs, majorly targeting the ATtinyX5 and ATtinyX7.
However, current `builder/main.py` will only upload when the board
is `digispark` varients.

Here we check the `upload.protocol` variable instead of `build.core`
to determine which uploader to use, so it's a lot easier to use
this powerful tiny uploader for different cores.