Add initial support for MicroCore ATtiny13
* Prepare for MCUdude MicroCore support
This commit PR a bug in the fuses.py script where the ATtiny13/A was expected to have an efuse.

* Use default programmer stk500v1 instead of arduinoisp
arduinoisp is NOT the same as arduinoasisp (formally known as stk500v1).
Most users doesn't own an arduinoisp (a rebranded usbtiny), but they rather use their Arduino and turn it into a programmer. stk500v1 is what we want.