Add SVD files for ATMega chips
Output of the https://github.com/Rahix/avr-device tool (all `*.svd.patched` files renamd to `*.svd`). 

Created as descripted in [community post](https://community.platformio.org/t/avr-simulator-as-aid-tool-for-debugging-code/13444/44?u=maxgerhardt)

The SVD files are not yet referenced in the board JSON files because the parser in VSCode throws an error and in CLion it can parse the file but cannot read out the registers (possible `avr-gdb` issue). Other issues will be opened for this.